refactor: 마이페이지 클래스 분리
