+++ src/main/java/com/ll/gooHaeYu/domain/member/member/controller/MemberController.java
@@ -1,12 +1,9 @@
 package com.ll.gooHaeYu.domain.member.member.controller;
 
-import com.ll.gooHaeYu.domain.application.application.dto.ApplicationDto;
-import com.ll.gooHaeYu.domain.application.application.service.ApplicationService;
-import com.ll.gooHaeYu.domain.jobPost.comment.dto.CommentDto;
-import com.ll.gooHaeYu.domain.jobPost.comment.service.CommentService;
-import com.ll.gooHaeYu.domain.jobPost.jobPost.dto.JobPostDto;
-import com.ll.gooHaeYu.domain.jobPost.jobPost.service.JobPostService;
-import com.ll.gooHaeYu.domain.member.member.dto.*;
+import com.ll.gooHaeYu.domain.member.member.dto.JoinForm;
+import com.ll.gooHaeYu.domain.member.member.dto.LoginForm;
+import com.ll.gooHaeYu.domain.member.member.dto.MemberDto;
+import com.ll.gooHaeYu.domain.member.member.dto.SocialProfileForm;
 import com.ll.gooHaeYu.domain.member.member.service.AuthenticationService;
 import com.ll.gooHaeYu.domain.member.member.service.MemberService;
 import com.ll.gooHaeYu.global.rsData.RsData;
@@ -23,7 +20,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
@@ -34,9 +30,6 @@ import java.util.stream.Stream;
 public class MemberController {
 
     private final MemberService memberService;
-    private final JobPostService jobPostService;
-    private final ApplicationService applicationService;
-    private final CommentService commentService;
     private final AuthenticationService authenticationService;
 
     @PostMapping("/join")
@@ -64,21 +57,6 @@ public class MemberController {
         return ResponseEntity.ok(responseMap);
     }
 
-    @GetMapping
-    @Operation(summary = "내 정보 조회")
-    public RsData<MemberDto> detailMember(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return RsData.of(memberService.findByUsername(memberDetails.getUsername()));
-    }
-
-    @PutMapping
-    @Operation(summary = "내 정보 수정")
-    public RsData<Void> modifyMember(@AuthenticationPrincipal MemberDetails memberDetails,
-                                             @Valid @RequestBody MemberForm.Modify form) {
-        memberService.modifyMember(memberDetails.getUsername(), form);
-
-        return RsData.of("204", "NO_CONTENT");
-    }
-
     @PutMapping("/social")
     @Operation(summary = "최초 소셜로그인 - 필수 회원정보 입력")
     public RsData<MemberDto> updateSocialMember(@AuthenticationPrincipal MemberDetails memberDetails,
@@ -88,28 +66,4 @@ public class MemberController {
         return RsData.of(updatedMember);
     }
 
-    @GetMapping("/myposts")
-    @Operation(summary = "내 공고 조회")
-    public RsData<List<JobPostDto>> detailMyPosts(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return RsData.of(jobPostService.findByUsername(memberDetails.getUsername()));
-    }
-
-
-    @GetMapping("/myapplications")
-    @Operation(summary = "내 지원서 조회")
-    public RsData<List<ApplicationDto>> detailMyApplications(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return RsData.of(applicationService.findByUsername(memberDetails.getUsername()));
-    }
-
-    @GetMapping("/mycomments")
-    @Operation(summary = "내 댓글 조회")
-    public RsData<List<CommentDto>> detailMyComments(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return RsData.of(commentService.findByUsername(memberDetails.getUsername()));
-    }
-
-    @GetMapping("/myinterest")
-    @Operation(summary = "내 관심 공고 조회")
-    public RsData<List<JobPostDto>> detailMyInterestingPosts(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return RsData.of(jobPostService.findByInterestAndUsername(memberDetails.getId()));
-    }
 }
